refactor: move DatasetCreateStatus to datasets.kafka
