+++ aineko/core/dataset.py
@@ -14,7 +14,7 @@ Example dataset configuration:
     ```
 """
 import abc
-from typing import Any, Dict, Generic, List, Optional, TypeVar
+from typing import Any, Dict, Generic, TypeVar
 
 from aineko.models.dataset_config_schema import DatasetConfig
 from aineko.utils.imports import import_from_string
@@ -33,58 +33,6 @@ class DatasetError(Exception):
     """
 
     pass
-
-
-class DatasetCreateStatus:
-    """Object representing staus of dataset creation.
-
-    Represents creation status of dataset (such as a kafka topic)
-    or its connections (such as producers and consumers linked
-    to the topic).
-
-    Can be used to ensure all datasets have been created.
-
-    Attributes:
-        dataset_name: Name of the dataset.
-        kafka_topic_to_future: Dictionary of kafka topics to futures.
-        status_list: List of status of dataset creation.
-
-    Args:
-        dataset_name: Name of the dataset.
-        kafka_topic_to_future: Dictionary of kafka topics to futures.
-        status_list: List of status of dataset creation.
-    """
-
-    def __init__(
-        self,
-        dataset_name: str,
-        kafka_topic_to_future: Optional[Dict[str, Any]] = None,
-        status_list: Optional[List[Any]] = None,
-    ):
-        """Creation status of dataset or its components."""
-        self.dataset_name = dataset_name
-        self.kafka_topic_to_future = kafka_topic_to_future
-        self.status_list = status_list
-
-    def done(self) -> bool:
-        """Return status of dataset creation.
-
-        For kafka topics, the status is represented by a dictionary
-        of kafka topics to futures. For kafka producers and consumers,
-        the status is represented by a list of status objects.
-
-        Returns:
-            True if all futures are done, otherwise False.
-        """
-        if not any([self.kafka_topic_to_future, self.status_list]):
-            return True
-        if self.kafka_topic_to_future:
-            return all(
-                future.done() for future in self.kafka_topic_to_future.values()
-            )
-        if self.status_list:
-            return all(status.done() for status in self.status_list)
-        return False
 
 
 class AbstractDataset(abc.ABC, Generic[T]):
