+++ aineko/datasets/kafka.py
@@ -26,11 +26,7 @@ from confluent_kafka.admin import AdminClient, NewTopic  # type: ignore
 from pydantic import BaseModel
 
 from aineko.config import AINEKO_CONFIG, DEFAULT_KAFKA_CONFIG
-from aineko.core.dataset import (
-    AbstractDataset,
-    DatasetCreateStatus,
-    DatasetError,
-)
+from aineko.core.dataset import AbstractDataset, DatasetError
 
 logger = logging.getLogger(__name__)
 
@@ -94,6 +90,58 @@ class ProducerParams(KafkaParams):
     producer_config: Dict[str, Any] = DEFAULT_KAFKA_CONFIG.get(
         "PRODUCER_CONFIG"
     )
+
+
+class DatasetCreateStatus:
+    """Object representing staus of dataset creation.
+
+    Represents creation status of dataset (such as a kafka topic)
+    or its connections (such as producers and consumers linked
+    to the topic).
+
+    Can be used to ensure all datasets have been created.
+
+    Attributes:
+        dataset_name: Name of the dataset.
+        kafka_topic_to_future: Dictionary of kafka topics to futures.
+        status_list: List of status of dataset creation.
+
+    Args:
+        dataset_name: Name of the dataset.
+        kafka_topic_to_future: Dictionary of kafka topics to futures.
+        status_list: List of status of dataset creation.
+    """
+
+    def __init__(
+        self,
+        dataset_name: str,
+        kafka_topic_to_future: Optional[Dict[str, Any]] = None,
+        status_list: Optional[List[Any]] = None,
+    ):
+        """Creation status of dataset or its components."""
+        self.dataset_name = dataset_name
+        self.kafka_topic_to_future = kafka_topic_to_future
+        self.status_list = status_list
+
+    def done(self) -> bool:
+        """Return status of dataset creation.
+
+        For kafka topics, the status is represented by a dictionary
+        of kafka topics to futures. For kafka producers and consumers,
+        the status is represented by a list of status objects.
+
+        Returns:
+            True if all futures are done, otherwise False.
+        """
+        if not any([self.kafka_topic_to_future, self.status_list]):
+            return True
+        if self.kafka_topic_to_future:
+            return all(
+                future.done() for future in self.kafka_topic_to_future.values()
+            )
+        if self.status_list:
+            return all(status.done() for status in self.status_list)
+        return False
 
 
 class KafkaDataset(AbstractDataset):
